+++ lib/store/helpers/add/index.ts
@@ -3,4 +3,5 @@ import { addSubItem } from "./addSubItem";
 import { addChoice } from "./addChoice";
 import { addLabeledChoice } from "./addLabeledChoice";
 import { addSubChoice } from "./addSubChoice";
-export { add, addSubItem, addChoice, addLabeledChoice, addSubChoice };
+import { duplicateElement } from "./duplicateElement";
+export { add, addSubItem, addChoice, addLabeledChoice, addSubChoice, duplicateElement };
